Moving stuff to utils
