+++ airflow/io/utils/stat.py
